Move FlashcardDialog, FlashcardWarningDialog into flashcards.
